Added functionality to login and logout.
Added LoginController, ServicePortalController and LogoutController.

LoginController accesses the DB to login the user.

ServicePortalController redirects the user to userPortal.jsp or staffPortal.jsp depending on their ROLE.

LogoutController when accessed invalidates the session and redirects back to the login page.

Also corrected some errorrs in the DB script. Accidentally had a comma rather than a period for the emails.